checkpoint  testing
feat: aws credentials bug is working.

https://gloo-global.slack.com/archives/C03KV1PJ6EM/p1736215459393779

new logic if we don't set any of the credentials manually in the
clients.baml, playground, or client registry then we will fallback to
use the environment variables. Otherwise, we have to set all the
environment variables in order to take them in.

remove runtest