feat: update server / client output for react generations
fix: nextjs-plugin issue where we were trying to include baml as a peer
dependency when it was not needed. This was causing the next.js app to
try to import multiple files of the *.node at compile time and causing
next.js to fail to compile.

in-progress: we are currently having an issue where we can't load baml
in a server action:

```
 at Object.__webpack_require__ [as require] (.next/server/webpack-runtime.js:33:42)
  357 |     //  - The user may need to bundle the correct files
  358 |     //  - The user may need to re-install node_modules to get new packages
> 359 |     throw new Error('Failed to load native binding', { cause: loadErrors })
      |          ^
  360 |   }
  361 |   throw new Error(`Failed to load native binding`)
  362 | } {
  digest: '857538051',
  [cause]: [Array]
  ```